+++ src/Tests/Unit/Public/Find-BingTimeZone.Tests.ps1
@@ -43,7 +43,7 @@ InModuleScope 'pwshPlaces' {
                     }
                 } #endMock
                 Find-BingTimeZone -PointLatitude 29.70 -PointLongitude -98.11 -BingMapsAPIKey $bingAPIKey
-                Assert-MockCalled -CommandName Write-Warning -Times 1
+                Should -Invoke -CommandName Write-Warning -Times 1
                 Assert-VerifiableMock
             } #it
 
@@ -67,7 +67,7 @@ InModuleScope 'pwshPlaces' {
                     BingMapsAPIKey = $bingAPIKey
                 }
                 Find-BingTimeZone @findBingTimeZoneSplat
-                Assert-MockCalled -CommandName Write-Warning -Times 1
+                Should -Invoke -CommandName Write-Warning -Times 1
                 Assert-VerifiableMock
             } #it