+++ src/Tests/Unit/Public/Invoke-GMapGeoCode.Tests.ps1
@@ -43,7 +43,7 @@ InModuleScope 'pwshPlaces' {
                     }
                 } #endMock
                 Invoke-GMapGeoCode -Address '24-593 Federation Drive, San Francisco, CA' -GoogleAPIKey $googleAPIKey
-                Assert-MockCalled -CommandName Write-Warning -Times 1
+                Should -Invoke -CommandName Write-Warning -Times 1
                 Assert-VerifiableMock
             } #it