+++ docs/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
         - updated actions from `v2` to `v3`
     - Moved `CHANGELOG.md` from `.github` to `docs`
     - `actions_bootstrap.ps1` - bumped module versions to latest
+    - Removed all test case uses of `Assert-MockCalled`
 - Misc
     - Updated `settings.json` for tab requirements to support Readthedocs
     - Added `SECURITY.md`